requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
bootstrap4==0.1.0
certifi==2021.10.8
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.1
django-stdimage==5.3.0
gunicorn==20.1.0
Pillow==9.0.0
progressbar2==4.0.0
psycopg-binary==3.0.8
psycopg2-binary==2.9.3
python-utils==3.1.0
pytz==2021.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
static3==0.7.0
